--- modules/database-mysql/impl/src/main/java/eu/cloudnetservice/modules/sql/impl/sqlite/SQLiteDatabaseProvider.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2019-present CloudNetService team & contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package eu.cloudnetservice.modules.sql.impl.sqlite;
-
-import eu.cloudnetservice.modules.mysql.config.JooqConfigurationEntry;
-import eu.cloudnetservice.modules.sql.impl.JooqDatabase;
-import eu.cloudnetservice.modules.sql.impl.JooqProvider;
-import eu.cloudnetservice.node.database.LocalDatabase;
-import lombok.NonNull;
-
-public class SQLiteDatabaseProvider extends JooqProvider {
-
-  protected SQLiteDatabaseProvider(@NonNull JooqConfigurationEntry config) {
-    super((_, _) -> {}, config);
-  }
-
-  @Override
-  public @NonNull LocalDatabase database(@NonNull String name) {
-    return new JooqDatabase(name, this, super.dslContext);
-  }
-}
