+++ modules/database-sql/impl/build.gradle.kts
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2019-present CloudNetService team & contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import eu.cloudnetservice.cloudnet.gradle.util.Files
+
+plugins {
+  id("cloudnet-modules")
+  id("cloudnet-publish")
+  alias(libs.plugins.shadow)
+}
+
+dependencies {
+  moduleLibrary(libs.bundles.sql) {
+    exclude("com.google.protobuf")
+  }
+
+  compileOnly(libs.caffeine)
+  compileOnlyApi(projects.node.nodeImpl)
+  api(projects.modules.databaseSql.databaseSqlApi)
+}
+
+tasks.shadowJar.configure {
+  archiveFileName = Files.databaseSql
+}
+
+moduleJson {
+  author = "CloudNetService"
+  name = "CloudNet-Database-SQL"
+  main = "eu.cloudnetservice.modules.sql.impl.CloudNetSQLModule"
+  description = "CloudNet extension, which includes the database support for MySQL, MariaDB, PostgreSQL and SQLite."
+  minJavaVersionId = JavaVersion.VERSION_11
+  runtimeModule = true
+  storesSensitiveData = true
+}
