+++ modules/database-sql/impl/src/test/java/eu/cloudnetservice/modules/sql/impl/SQLiteDatabaseTest.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2019-present CloudNetService team & contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package eu.cloudnetservice.modules.sql.impl;
+
+import eu.cloudnetservice.modules.sql.config.DatabaseType;
+import eu.cloudnetservice.modules.sql.config.JooqConfigurationEntry;
+import eu.cloudnetservice.modules.sql.impl.junit.EnableServicesInject;
+import org.junit.jupiter.api.BeforeAll;
+
+@EnableServicesInject
+public class SQLiteDatabaseTest extends SQLDatabaseTest {
+
+  @BeforeAll
+  static void setup() throws Exception {
+    var config = new JooqConfigurationEntry(
+      DatabaseType.SQLITE,
+      "sqlite",
+      "cn_testing",
+      null,
+      null,
+      null,
+      "jdbc:sqlite::memory:");
+    databaseProvider = new JooqProvider(config);
+    databaseProvider.init();
+  }
+}
